remove creating attendance from swagger and api AND get attendance by subject id is added
